pyproject.toml
[tool.poetry]
name = "visa-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "visa_app"}]

[tool.poetry.dependencies]
python = "^3.11"
selenium = "^4.22.0"
webdriver-manager = "^4.0.1"
pygame = "^2.6.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
